+++ flutter/sherpa_onnx/lib/sherpa_onnx.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 import 'dart:ffi';
 
 export 'src/feature_config.dart';
+export 'src/keyword_spotter.dart';
 export 'src/offline_recognizer.dart';
 export 'src/offline_stream.dart';
 export 'src/online_recognizer.dart';